[ATK] accessibility/aria-current-global-attribute.html has been failed since r198303 https://bugs.webkit.org/show_bug.cgi?id=155935
Reviewed by Chris Fleizach.

Source/WebCore:

In ATK, the text of span elements is typically exposed through the parent
element, unless the span has some attribute requiring inclusion in the
accessibility tree. We were not checking for the presence of global ARIA
attributes (including aria-current), other than describedby and role.
The fix is to also check AccessibilityObject::supportsARIAAttributes().

No new tests. This issue was caught as a result of a failing test.

* accessibility/atk/AccessibilityObjectAtk.cpp:
(WebCore::AccessibilityObject::accessibilityPlatformIncludesObject):

LayoutTests:

* platform/efl/TestExpectations: Unskipped the failing test.


Identifier: 175284@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@200244 268f45cc-cd09-0410-ab3c-d52691b4dbfc