Unreviewed GTK+ gardening. Update drag and drop tests bug reference.
We were using a fixed bug, because drag and drop was already implemented in WebKit2. However, some tests still
fail because some drag and drop events are not emitted when using synthetic events like even sender does. So,
update the bug reference to https://bugs.webkit.org/show_bug.cgi?id=157179 and add two more tests that are
timing out in the bots for the same reason.

* platform/gtk/TestExpectations:

Identifier: 175400@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@200366 268f45cc-cd09-0410-ab3c-d52691b4dbfc