Stop traversing at the container block when computing RTL inline static distance. https://bugs.webkit.org/show_bug.cgi?id=157349 <rdar://problem/25994087>
Reviewed by David Hyatt.

When computing the inline static distance for a child renderer, we start at its enclosing box
and traverse up all the way to the container block.
However when the enclosing box is the ancestor of the container block, we
should just bail out right away since there's no container to use to adjust the position.

Source/WebCore:

Test: fast/multicol/positioned-rtl-column-crash.html

* rendering/RenderBox.cpp:
(WebCore::computeInlineStaticDistance):

LayoutTests:

* fast/multicol/positioned-rtl-column-crash-expected.txt: Added.
* fast/multicol/positioned-rtl-column-crash.html: Added.

Identifier: 175523@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@200492 268f45cc-cd09-0410-ab3c-d52691b4dbfc