4 and 8 digit color attributes should use the crazypants algorithm, not the CSS one. https://bugs.webkit.org/show_bug.cgi?id=157402 <rdar://problem/26131142>
Reviewed by Simon Fraser.

Source/WebCore:

Noel and Tab from Google alerted me to the fact I broke some legacy
HTML color attributes when I added support for 4 and 8 digit hex colors
in CSS. The fix is to favor the "crazy" parsing unless we know it is a 3 or 6
digit hex value (in which case we try the CSS algorithm first).

Covered by reverting an existing test.

* html/HTMLElement.cpp:
(WebCore::HTMLElement::addHTMLColorToStyle):

LayoutTests:

Restore the version that was there before I added support for
4 and 8 digit hex colors in CSS.

* fast/dom/attribute-legacy-colors-expected.txt:
* fast/dom/script-tests/attribute-legacy-colors.js:

Identifier: 175532@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@200501 268f45cc-cd09-0410-ab3c-d52691b4dbfc