REGRESSION(r196222): [AX][GTK] accessibility/gtk/caret-offsets.html failing https://bugs.webkit.org/show_bug.cgi?id=153956
Reviewed by Chris Fleizach.

Source/WebCore:

The reason the test began failing is that it was checking the new caret offset
synchronously. For most of the test cases, this was not a problem. But when the
caret was moved out of a focused link, the focus change (and associated repainting)
delayed the caret-moved event long enough to cause the associated test case to fail.
The test now uses shouldBecomeEqualToString() instead of shouldBeEqualToString().

The test also had a supposedly-correct expectation which was wrong: When moving the
caret to a valid accessible offset, the caret-moved event should be for that offset.
This was not the case for the list item test case because emitTextSelectionChange()
was not adjusting the offset for the RenderListMarker, the text of which is exposed
as part of the ATK_ROLE_LIST_ITEM object. This bug was also fixed and the test case
updated accordingly.

No new tests are needed. The previously-failing test is now passing.

* editing/atk/FrameSelectionAtk.cpp:
(WebCore::emitTextSelectionChange):

LayoutTests:

Change the test to run asynchronously and update the expectations.

* accessibility/gtk/caret-offsets-expected.txt: Updated.
* accessibility/gtk/caret-offsets.html: Updated.
* platform/gtk/TestExpectations: Unskipped the previously-failing test.


Identifier: 175573@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@200542 268f45cc-cd09-0410-ab3c-d52691b4dbfc